(+) Optimize RecyclerView Item (+) Debugging Database Receiver (+) Redesign NEws Item (+) Redesign Province Item (+) Redesign News Loading

+++ app/src/main/java/com/github/nothing2512/anticorona/services/DatabaseService.kt
@@ -33,7 +33,7 @@ class DatabaseService {
      * Deleting data in database
      * @param context
      */
-     fun delete(context: Context) {
+    fun delete(context: Context, all: Boolean = false) {
 
         /**
          * Providing Database
@@ -56,7 +56,7 @@ class DatabaseService {
         casesDao.deleteCountries()
 
         newsDao.delete()
-        faqsDao.delete()
+        if (all) faqsDao.delete()
     }
 
     /**
